// app/src/app/api/model/lifecyclePhase.ts
/**
 * TOSCAna
 * To be Done!
 *
 * OpenAPI spec version: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface LifecyclePhase {
    /**
     * the name of this execution phase
     */
    name: string;
    /**
     * the current state of the phase. Must be one of { PENDING, SKIPPING, EXECUTING, SKIPPED, DONE, ERROR }
     */
    state: LifecyclePhase.StateEnum;
}
export namespace LifecyclePhase {
    export type StateEnum = 'PENDING' | 'SKIPPING' | 'EXECUTING' | 'DONE' | 'SKIPPED' | 'FAILED';
    export const StateEnum = {
        PENDING: 'PENDING' as StateEnum,
        SKIPPING: 'SKIPPING' as StateEnum,
        EXECUTING: 'EXECUTING' as StateEnum,
        DONE: 'DONE' as StateEnum,
        SKIPPED: 'SKIPPED' as StateEnum,
        FAILED: 'FAILED' as StateEnum
    }
}